unify grpc and jsonrpc client device disabled tests
